+++ packages/ranger-ui/tsup.config.ts
@@ -19,15 +19,6 @@ export default defineConfig((options: Options) => {
     terserOptions: {
       compress: true
     },
-    esbuildPlugins: [
-      replacePlugin({
-        values: {
-          // 使用正则表达式替换掉模板字符串中的换行符
-          '\\n': '',
-          // 如果需要移除所有空白符（包括空格），可以使用以下正则表达式
-          '\\s+': ''
-        }
-      })
-    ]
+    esbuildPlugins: [replacePlugin()]
   }
 })